[NFC] Split the SIL type substitution code into its own file
